Always writes rent_epoch to storage as u64::MAX
